multi-restaurant-delivery/backend/recommender/requirements.txt
fastapi==0.95.0
redis==4.5.5
psycopg[binary]==3.1.18
numpy==1.24.4
scikit-learn==1.2.2
kafka-python==2.0.2
